Fix: 'adjust_inventory_host' should generate template variables
The 'adjust_inventory_host' is used to add group vars & co to node
data. However, it's also used to prep the node data for config template
generation, and in that case we need variables that are usually
computed in Ansible playbooks.

The new 'template_vars' parameter triggers generation of those
variables (inventory_hostname, netlab_interfaces...)